Move consts package to internal as preparations for v1.0.0
