+++ internal/js/bundle_test.go
@@ -20,11 +20,11 @@ import (
 	"gopkg.in/guregu/null.v3"
 
 	"go.k6.io/k6/internal/build"
+	"go.k6.io/k6/internal/lib/consts"
 	"go.k6.io/k6/internal/lib/testutils"
 	"go.k6.io/k6/internal/loader"
 	"go.k6.io/k6/internal/usage"
 	"go.k6.io/k6/lib"
-	"go.k6.io/k6/lib/consts"
 	"go.k6.io/k6/lib/fsext"
 	"go.k6.io/k6/lib/types"
 	"go.k6.io/k6/metrics"
