+++ internal/js/timeout_error_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"go.k6.io/k6/lib/consts"
+	"go.k6.io/k6/internal/lib/consts"
 )
 
 func TestTimeoutError(t *testing.T) {
